Formatting

+++ scripts/iiif.py
@@ -1,6 +1,6 @@
 import sys
 
-sys.path.insert(0, './classes')
+sys.path.insert(0, "./classes")
 import argparse
 import json
 import logging
@@ -17,19 +17,26 @@ from item import BUCKET, COLLECTIONS, Item
 from tqdm import tqdm
 from tqdm.contrib.logging import logging_redirect_tqdm
 
-from helpers import load_xls, fast_upload, logger, session, upload_folder_to_s3 #, invalidate_cache
+from helpers import (
+    load_xls,
+    fast_upload,
+    logger,
+    session,
+    upload_folder_to_s3,
+)  # , invalidate_cache
 
 load_dotenv(override=True)
+
 
 def get_items(metadata, vocabulary, mode):
 
     os.makedirs(COLLECTIONS, exist_ok=True)
 
     # list all collection names
-    collections = metadata["Collections"].dropna().str.split("\|").explode().unique()
+    collections = metadata["Collection"].dropna().str.split("\|").explode().unique()
     # download collection(s)
     for name in collections:
-        collection_path = f"{COLLECTIONS}{name.lower()}.json"
+        collection_path = COLLECTIONS + name.lower() + ".json"
         response = session.get(BUCKET + collection_path)
         if response.status_code == 200:
             with open(collection_path, "w") as f:
@@ -55,10 +62,11 @@ def get_items(metadata, vocabulary, mode):
     #         to_process = metadata
     # else:
     # to_process = metadata
-        
+
     ##logger.debug(f"Processing {len(to_process)} items")
 
     return [Item(id, row, vocabulary) for id, row in metadata.fillna("").iterrows()]
+
 
 if __name__ == "__main__":
 
@@ -67,9 +75,7 @@ if __name__ == "__main__":
     parser.add_argument(
         "--mode", "-m", help="run mode", choices=["test", "prod"], default="test"
     )
-    parser.add_argument(
-        "--index", "-i", help="index to run"
-    )
+    parser.add_argument("--index", "-i", help="index to run", default=0)
     args = parser.parse_args()
 
     if os.path.exists("/iiif"):
@@ -86,31 +92,33 @@ if __name__ == "__main__":
         metadata = metadata.loc[metadata["Status"] == "In imagineRio"]
 
     # filter columns
-    metadata = metadata[[
-        "Document ID",
-        "Rights",
-        "Provider",
-        "Collections",
-        "Type",
-        "Material",
-        "Fabrication Method",
-        "Media URL",
-        "Creator",
-        "Title",
-        "Description (Portuguese)",
-        "Description (English)",
-        "Date",
-        "First Year",
-        "Last Year",
-        "Document URL",
-        "Required Statement",
-        "Wikidata ID",
-        "Smapshot ID",
-        "Depicts",
-        "Width",
-        "Height"
-    ]]
-    
+    metadata = metadata[
+        [
+            "Document ID",
+            "Rights",
+            "Provider",
+            "Collection",
+            "Type",
+            "Material",
+            "Fabrication Method",
+            "Media URL",
+            "Creator",
+            "Title",
+            "Description (Portuguese)",
+            "Description (English)",
+            "Date",
+            "First Year",
+            "Last Year",
+            "Document URL",
+            "Required Statement",
+            "Wikidata ID",
+            "Smapshot ID",
+            "Depicts",
+            "Width",
+            "Height",
+        ]
+    ]
+
     items = get_items(metadata, vocabulary, args.mode)
 
     with logging_redirect_tqdm():
@@ -120,20 +128,35 @@ if __name__ == "__main__":
             main_pbar.set_postfix_str(str(item._id))
             success = item.write_manifest()
 
-            if success: # TODO item.upload_and_remove()
-                tiles = [os.path.join(root, file) for root, _, files in os.walk(item._base_path) for file in files]
+            if success and args.mode == "prod":  # TODO item.upload_and_remove()
+                tiles = [
+                    os.path.join(root, file)
+                    for root, _, files in os.walk(item._base_path)
+                    for file in files
+                ]
                 total_size = sum([os.stat(f).st_size for f in tiles])
-                with tqdm(tiles, desc='upload', ncols=60, total=total_size, 
-                unit='B', unit_scale=1, leave=False) as upload_pbar:
-                    fast_upload(boto3.Session(), "imaginerio-images", tiles, upload_pbar.update)
-                #tqdm.write(f"Upload complete, removing folder {item._base_path}")
+                with tqdm(
+                    tiles,
+                    desc="upload",
+                    ncols=60,
+                    total=total_size,
+                    unit="B",
+                    unit_scale=1,
+                    leave=False,
+                ) as upload_pbar:
+                    fast_upload(
+                        boto3.Session(), "imaginerio-images", tiles, upload_pbar.update
+                    )
+                # tqdm.write(f"Upload complete, removing folder {item._base_path}")
                 shutil.rmtree(os.path.abspath(item._base_path))
             else:
                 continue
-            
-        fast_upload(boto3.Session(), "imaginerio-images", [
-            os.path.relpath(file) for file in os.listdir(COLLECTIONS)
-            ])
-        #upload_folder_to_s3(COLLECTIONS, mode=args.mode)
+        if args.mode == "prod":
+            fast_upload(
+                boto3.Session(),
+                "imaginerio-images",
+                [os.path.relpath(file) for file in os.listdir(COLLECTIONS)],
+            )
+            # upload_folder_to_s3(COLLECTIONS, mode=args.mode)
 
-    #invalidate_cache("/*")
+    # invalidate_cache("/*")
